requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.16.0
contextlib2==21.6.0
cryptography==42.0.8
Django==5.0.6
django-cors-headers==4.4.0
django-ninja==1.2.0
django-ninja-extra==0.21.0
django-ninja-jwt==5.3.1
fire==0.6.0
injector==0.21.0
markdown-it-py==3.0.0
mdurl==0.1.2
pycparser==2.22
pydantic==2.8.0
pydantic_core==2.20.0
Pygments==2.18.0
PyJWT==2.8.0
PyYAML==6.0.1
rav==0.0.9
rich==13.7.1
six==1.16.0
sqlparse==0.5.0
termcolor==2.4.0
typing_extensions==4.12.2
tzdata==2024.1
